modified all files that referenced StationInfo. Renamed Station_Info to station_info and deleted original station_info folder
