--- src/use_case/StationInfo/StationInfoInputBoundary.java
@@ -1,5 +1,0 @@
-package use_case.StationInfo;
-
-public interface StationInfoInputBoundary {
-    void execute(StationInfoInputData stationInfoInputData);
-}
