--- src/use_case/StationInfo/StationInfoInputData.java
@@ -1,9 +1,0 @@
-package use_case.StationInfo;
-
-public class StationInfoInputData {
-    final private String stationName;
-
-    public StationInfoInputData(String stationName) {this.stationName = stationName;}
-
-    String getStationName() {return stationName;}
-}
