--- src/use_case/StationInfo/StationInfoOutputBoundary.java
@@ -1,7 +1,0 @@
-package use_case.StationInfo;
-
-public interface StationInfoOutputBoundary {
-    void prepareSuccessView(StationInfoOutputData stationInfoOutputData);
-
-    void parepareFailView(String error);
-}
