+++ src/use_case/station_info/StationInfoInputBoundary.java
@@ -1,7 +1,5 @@
 package use_case.station_info;
 
-import use_case.search.SearchInputData;
-
 public interface StationInfoInputBoundary {
     void execute(StationInfoInputData stationInfoInputData);
 }
