+++ src/use_case/station_info/StationInfoOutputBoundary.java
@@ -1,4 +1,7 @@
 package use_case.station_info;
 
 public interface StationInfoOutputBoundary {
+    void prepareSuccessView(StationInfoOutputData stationInfoOutputData);
+
+    void parepareFailView(String error);
 }
