--- src/use_case/station_info/StationOutputData.java
@@ -1,4 +1,0 @@
-package use_case.station_info;
-
-public class StationOutputData {
-}
